git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@5843 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/fx.c
@@ -2633,7 +2633,7 @@ static MagickRealType FxEvaluateSubexpression(FxInfo *fx_info,
         {
           alpha=FxEvaluateSubexpression(fx_info,channel,x,y,expression+5,beta,
             exception);
-          return((MagickRealType) !!isnan(alpha));
+          return((MagickRealType) !!isnan((double) alpha));
         }
       if (LocaleCompare(expression,"i") == 0)
         return(FxGetSymbol(fx_info,channel,x,y,expression,exception));
